Re-organize vdex classes
